Refactor fix bug
Removed unused vars
Moved components to component folder
Fixed bug on MyPage
