+++ src/components/AllGames.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import { getGames, formatWinner, formatDate } from './util';
+import { getGames } from '../util';
 import Games from './Games';
-import './css/Games.css';
+import '../css/Games.css';
 
 export default class AllGames extends Component {
