+++ src/components/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Main from './Main';
 import Header from './Header';
 import 'spectre.css/docs/dist/spectre.css';
-import './App.css';
+import '../css/App.css';
 import 'react-toastify/dist/ReactToastify.min.css';
 import Toast from './Toast';
